rename buyan_logger file name
