+++ app/services/bunyan_logger.js
